INTPO002PD-2558 - Cannot register new user - error 500 due to the duplicate key

+++ src/main/resources/db/migration/V20220602161100__add_cart_products_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS cart_products
 (
-    cart_id     BIGSERIAL NOT NULL,
-    products_id BIGSERIAL NOT NULL
+    cart_id     INTEGER NOT NULL,
+    products_id INTEGER NOT NULL
 );
